Update users.js
Start Predict is a project that integrates data from start.gg, allowing users to make predictions for any tournament. It offers pre- and post-match statistics, provides insights into results, and even enables users to download their predictions. With a range of features designed for in-depth analysis, it enhances the tournament experience for competitive gamers.